+++ sample_app/lib/pages/new_group_chat_screen.dart
@@ -1,5 +1,3 @@
-import 'dart:async';
-
 import 'package:flutter/material.dart';
 import 'package:go_router/go_router.dart';
 import 'package:sample_app/routes/routes.dart';
@@ -23,8 +21,6 @@ class _NewGroupChatScreenState extends State<NewGroupChatScreen> {
 
   bool _isSearchActive = false;
 
-  Timer? _debounce;
-
   late final userListController = StreamUserListController(
     client: StreamChat.of(context).client,
     sort: [const SortOption.asc('name')],
@@ -35,20 +31,18 @@ class _NewGroupChatScreenState extends State<NewGroupChatScreen> {
   );
 
   void _userNameListener() {
-    if (_debounce?.isActive ?? false) _debounce!.cancel();
-    _debounce = Timer(const Duration(milliseconds: 350), () {
-      if (mounted) {
-        setState(() {
-          _userNameQuery = _controller.text;
-          _isSearchActive = _userNameQuery.isNotEmpty;
-        });
-        userListController.filter = Filter.and([
-          if (_userNameQuery.isNotEmpty) Filter.autoComplete('name', _userNameQuery),
-          Filter.notEqual('id', StreamChat.of(context).currentUser!.id),
-        ]);
-        userListController.doInitialLoad();
-      }
+    setState(() {
+      _userNameQuery = _controller.text;
+      _isSearchActive = _userNameQuery.isNotEmpty;
     });
+
+    return userListController.search(
+      _userNameQuery,
+      filter: Filter.and([
+        if (_userNameQuery.isNotEmpty) Filter.autoComplete('name', _userNameQuery),
+        Filter.notEqual('id', StreamChat.of(context).currentUser!.id),
+      ]),
+    );
   }
 
   @override